move `Node` and `Network` out of `yapapi.rest` for the sake of better separation between the REST API and the high-level API
